chan_crawler/requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
Pebble==5.0.7
psycopg2-binary==2.9.6
pydantic==2.10.3
pydantic_core==2.27.1
pyfaktory==0.2.8
python-dotenv==1.0.1
requests==2.32.3
rfc3339-validator==0.1.4
six==1.17.0
typing_extensions==4.12.2
urllib3==2.2.3
